Here's my little function I've been working on. If anyone has R, they can download this script and the test.csv file to check it out.
I need to figure out a better way to weight categories, though.

What does it do? Here it takes a dataset on animal outcomes. You can select, say, all the animals that were put down. Then it will got through columns that you select related to that data.

There are categories related to sex, breed, status etc. It will count each column and get frequencies related to each category. How many pit bulls, how many cats/dogs etc. If it was a common characteristic, then rows containing things that were more frequent will be weighted higher and thus be sorted to the top. Those animals with higher weights will be "more likely" to be put down.

There are problems with using sums though. Just because there are more cats that are put down doesn't mean that a specific cat is more likely to be put down. It might just mean that there are a lot of cats around. That's what I've got to change next.